[axes-math-model] fix a circular dependency (via SelectAndDragByOptions of motion-stage) by splitting out the models from axes-math
