+++ lib/js/components/layouts/type-tools-grid.mjs
@@ -63,7 +63,10 @@ import {
 import {
     AxesMathAxisLocationValueModel
   , AxesMathAxisLocationValuesModel
-  , UIAxesMathLocationValue
+} from '../axes-math-models.mjs';
+
+import {
+    UIAxesMathLocationValue
   , UIAxesMathLocationValues
 } from '../axes-math.mjs';
