+++ lib/js/components/type-dimension-fundamentals.mjs
@@ -23,7 +23,7 @@ import {
 import {
     AxesMathAxisLocationValueModel
   , AxesMathAxisLocationValuesModel
-} from './axes-math.mjs';
+} from './axes-math-models.mjs';
 
 // FIXME: Consider explicitly including the term "discrete" in the name
 // and/or documentation of this module. The term "dimension" alone may
