+++ lib/js/components/type-driven-ui.mjs
@@ -102,7 +102,10 @@ import {
 
 import {
     AxesMathAxisLocationsModel
-  , UIAxesMathLocation
+} from './axes-math-models.mjs';
+
+import {
+    UIAxesMathLocation
 } from './axes-math.mjs';
 
 import {
