+++ lib/js/components/type-spec-models.mjs
@@ -49,7 +49,7 @@ import {
 
 import {
     AxesMathAxisLocationsModel
-} from './axes-math.mjs';
+} from './axes-math-models.mjs';
 
 import {
     GENERIC
